backend/requirements.txt
Django>=3.2
djangorestframework
Pillow
django-cors-headers
